feat(devtools): add draft exploration
